rolend 8-8 19:06

+++ src/main/java/com/sbaitproject/roadwith/region/vo/Zone.java
@@ -2,7 +2,7 @@ package com.sbaitproject.roadwith.region.vo;
 
 public class Zone {
 	private String sigu;
-	private int zoneNumber;
+	private double zoneNumber;
 	
 	public String getSigu() {
 		return sigu;
@@ -10,10 +10,10 @@ public class Zone {
 	public void setSigu(String sigu) {
 		this.sigu = sigu;
 	}
-	public int getZoneNumber() {
+	public double getZoneNumber() {
 		return zoneNumber;
 	}
-	public void setZoneNumber(int zoneNumber) {
+	public void setZoneNumber(double zoneNumber) {
 		this.zoneNumber = zoneNumber;
 	}
 }
